### 2023-02-16 #5159 @editorialbot
[PRE REVIEW]: Hypercomplex: abstract & fast header-only C++ template library for lattice-based cryptosystems in high-dimensional algebras
**Submitting author:** @AngryMaciek (<a href="http://orcid.org/0000-0003-1361-7301">Maciej Bak</a>)
**Repository:** https://github.com/AngryMaciek/hypercomplex
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
