### 2023-02-16 #5159 @AngryMaciek
Hi @danielskatz : strange, the CI Action with paper draft compilation works fine in the repo...  Well, I am not currently affiliated with any institution and I highly doubt that the previous ones would allow me to use their for a personal project. Is there any solution you would suggest in this situation?
