### 2023-02-16 #5159 @danielskatz
@AngryMaciek - as the example paper shows for author 3, please use "Independent Researcher, Country" in this case
