### 2023-02-16 #5159 @AngryMaciek
Hi @olexandr-konovalov,

Thank you for agreeing to oversee this process.

The story behind this work is that I have already submitted once to _JOSS_, [here is the revision](https://github.com/openjournals/joss-reviews/issues/3096) (so this is a re-submission now). At that point in time I had a pretty solid framework for operations on hypercomplex numbers. However, that was too abstract and I could not provide a concrete example of a use case. I have withdrawn the submission and went back to my basement to study cryptography for three years. Now I come back with a library, which actually core feature is the support for cryptosystems based on polynomial quotient rings (like [NTRU](https://en.wikipedia.org/wiki/NTRUEncrypt)) but generalised on higher dimensions. Up to my knowledge such systems are not available yet, moreover - here we have a working implementation.

As we already went through the revisions once I have elevated the repository to a high standard in terms of scientific software engineering. What I am not 100% certain about is the novel cryptographic part. That is: the ecnryption/decryption procedures work, one may check in the unit test CI workflow (GitHub Actions). However, when it comes to the revision - I would kindly ask for mathematicians to double-check every character in the `Cryptographic Application` section; it may require some polish 😅

Kind Regards,
Maciek
