### 2023-02-16 #5159 @olexandr-konovalov
Hi @AngryMaciek, thank you for the background details!

Do you have any suggestions for potential reviewers? If so, you can mention them here in this thread (without tagging them with an `@`, please).

In addition, you can check our [list of potential reviewers](https://bit.ly/joss-reviewers). They have already agreed to review for JOSS, and some of them may be suitable for your submission (please start the search from the bottom of the list).
