### 2023-02-17 #5159 @olexandr-konovalov
@ceb8 and @kpeeters - would one or both of you be able to review this submission for JOSS, please?

If not, could you possibly recommend someone whom we may approach regarding this?
