### 2023-03-10 #5159 @olexandr-konovalov
Hi @defeo, is there a chance that you would be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
