### 2023-03-10 #5159 @olexandr-konovalov
Thank you @defeo! Hi @gregorseiler and @lducas - would one (or both) of you be willing to review this submission for JOSS? If you haven't reviewed for JOSS before, here we have the [Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
