### 2023-03-30 AngryMaciek/hypercomplex#35 @ludopulles
During compilation I also ran into this problem: https://github.com/catchorg/Catch2/issues/2421.
