### 2023-04-05 AngryMaciek/hypercomplex#35 @vissarion
I have a similar compilation error (and a few warnings) with `clang-14`
```
mkdir ../.test/unit/hypercomplex; \
cp Hypercomplex.hpp ../.test/unit/hypercomplex/Hypercomplex.hpp; \
cp Polynomial.hpp ../.test/unit/hypercomplex/Polynomial.hpp; \
cd ../.test/unit; \
clang++ -O0 -Wall --std=c++17 -o test test.cpp -lmpfr -lgmp; \
./test -d yes -w NoAssertions --use-colour yes --benchmark-samples 100 --benchmark-resamples 100000; \
rm -rf hypercomplex test
In file included from test.cpp:16:
./catch.hpp:10818:34: error: constexpr variable 'sigStackSize' must be initialized by a constant expression
    static constexpr std::size_t sigStackSize = 32768 >= MINSIGSTKSZ ? 32768 : MINSIGSTKSZ;
                                 ^              ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
./catch.hpp:10818:58: note: non-constexpr function 'sysconf' cannot be used in a constant expression
    static constexpr std::size_t sigStackSize = 32768 >= MINSIGSTKSZ ? 32768 : MINSIGSTKSZ;
                                                         ^
/usr/include/x86_64-linux-gnu/bits/sigstksz.h:32:22: note: expanded from macro 'MINSIGSTKSZ'
# define MINSIGSTKSZ SIGSTKSZ
                     ^
/usr/include/x86_64-linux-gnu/bits/sigstksz.h:28:19: note: expanded from macro 'SIGSTKSZ'
# define SIGSTKSZ sysconf (_SC_SIGSTKSZ)
                  ^
/usr/include/unistd.h:640:17: note: declared here
extern long int sysconf (int __name) __THROW;
                ^
In file included from test.cpp:16:
./catch.hpp:10877:33: error: variable length array declaration not allowed at file scope
    char FatalConditionHandler::altStackMem[sigStackSize] = {};
                                ^           ~~~~~~~~~~~~
test.cpp:244:12: warning: explicitly assigning value of variable of type 'Hypercomplex<TestType, dim4>' to itself [-Wself-assign-overloaded]
        h1 = h1;
        ~~ ^ ~~
test.cpp:937:16: warning: explicitly assigning value of variable of type 'Hypercomplex<mpfr_t, dim4>' (aka 'Hypercomplex<__mpfr_struct[1], dim4>') to itself [-Wself-assign-overloaded]
            h1 = h1;
            ~~ ^ ~~
test.cpp:1457:12: warning: explicitly assigning value of variable of type 'Polynomial<deg>' to itself [-Wself-assign-overloaded]
        P1 = P1;
        ~~ ^ ~~
test.cpp:1886:16: warning: explicitly assigning value of variable of type 'Hypercomplex<Polynomial<MaxDeg>, dim>' to itself [-Wself-assign-overloaded]
            hA = hA;
            ~~ ^ ~~
4 warnings and 2 errors generated.
```
