### 2023-04-05 AngryMaciek/hypercomplex#36 @vissarion
JOSS: MPFR assertion failed
The following code
```
#include "hypercomplex/Hypercomplex.hpp"

int main(void){

  mpfr_t A[8];
  mpfr_init2(A[0], get_mpfr_precision());
  mpfr_init2(A[1], get_mpfr_precision());
  mpfr_init2(A[2], get_mpfr_precision());
  mpfr_init2(A[3], get_mpfr_precision());
  mpfr_init2(A[4], get_mpfr_precision());
  mpfr_init2(A[5], get_mpfr_precision());
  mpfr_init2(A[6], get_mpfr_precision());
  mpfr_init2(A[7], get_mpfr_precision());
  mpfr_set_d(A[0], 1.5, MPFR_RNDN);
  mpfr_set_d(A[1], 2.5, MPFR_RNDN);
  mpfr_set_d(A[2], 0.0, MPFR_RNDN);
  mpfr_set_d(A[3], -1.5, MPFR_RNDN);
  mpfr_set_d(A[4], 0.5, MPFR_RNDN);
  mpfr_set_d(A[5], -0.5, MPFR_RNDN);
  mpfr_set_d(A[6], -0.5, MPFR_RNDN);
  mpfr_set_d(A[7], -1.5, MPFR_RNDN);

  Hypercomplex<mpfr_t, 8> Hx(A);

  std::cout << "Hx^30 = ";
  mpfr_out_str(stdout, 10, 0, (Hx^30)[0], MPFR_RNDN);
  std::cout << std::endl;

  return 0;
}
```
compiled with `g++ --std=c++17 -lmpfr -lgmp ` (g++ version 12.2.0, mprf version 4.1.0) returns
```
../../src/init2.c:52: MPFR assertion failed: ((p) >= 1 && (p) <= ((mpfr_prec_t) ((((mpfr_uprec_t) -1) >> 1) - 256)))
Aborted (core dumped)
```

This is part of my review for JOSS https://github.com/openjournals/joss-reviews/issues/5272
