### 2023-04-15 AngryMaciek/hypercomplex#35 @AngryMaciek
@ludopulles

1. typo: template parameter should be $n+1$ for an array of size $n$
2. My bad, the test code:
```
            short int si = 3;
            unsigned short int usi = 3;
            int i = 3;
            unsigned int ui = 3;
            long int li = 3;
            unsigned long int uli = 3;
            long long int lli = 3;
            unsigned long long int ulli = 3;
            REQUIRE_NOTHROW(hA ^= si);
            REQUIRE_NOTHROW(hA ^= usi);
            REQUIRE_NOTHROW(hA ^= i);
            REQUIRE_NOTHROW(hA ^= ui);
            REQUIRE_NOTHROW(hA ^= li);
            REQUIRE_NOTHROW(hA ^= uli);
            REQUIRE_NOTHROW(hA ^= lli);
            REQUIRE_NOTHROW(hA ^= ulli);
```
is a series of power-assignments, ending up with `hA ^ 6561`; obviously there are limits for `long long` arithmetic.

Will be fixed in a new branch :ok:
