+++ .test/unit/test.cpp
@@ -1549,7 +1549,7 @@ TEST_CASE( "Polynomial: const objects", "[unit]" ) {
 
 TEST_CASE( "Polynomial: CenteredLift function", "[unit]" ) {
     //
-    int64_t coefficients_1[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+    int64_t coefficients_1[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
     Polynomial<10> P1(coefficients_1);
     CenteredLift(&P1, 12);
     REQUIRE( P1[0] == 0 );
@@ -1562,6 +1562,7 @@ TEST_CASE( "Polynomial: CenteredLift function", "[unit]" ) {
     REQUIRE( P1[7] == -5 );
     REQUIRE( P1[8] == -4 );
     REQUIRE( P1[9] == -3 );
+    REQUIRE( P1[10] == -2 );
 }
 
 TEST_CASE( "Polynomial: RingInverse function", "[unit]" ) {
@@ -1949,7 +1950,7 @@ TEST_CASE( "Hypercomplex: Polynomial lib test", "[unit]" ) {
             REQUIRE( hA[2] == polynomial3 % 2 );
             REQUIRE( hA[3] == polynomial4 % 2 );
         }
-
+        
         SECTION( "Multiplication operator" ) {
             Hypercomplex<Polynomial<MaxDeg>, dim> h = hA * hB;
             REQUIRE( h[0][0] == -43 );
@@ -1973,7 +1974,7 @@ TEST_CASE( "Hypercomplex: Polynomial lib test", "[unit]" ) {
             REQUIRE( h[3][3] == 32 );
             REQUIRE( h[3][4] == 13 );
         }
-
+        
         SECTION( "Power operator" ) {
             REQUIRE_THROWS_AS(hA ^ 0, std::invalid_argument);
             REQUIRE_NOTHROW(hA ^ 1);
@@ -2016,7 +2017,7 @@ TEST_CASE( "Hypercomplex: Polynomial lib test", "[unit]" ) {
             REQUIRE_NOTHROW(hA ^ lli);
             REQUIRE_NOTHROW(hA ^ ulli);
         }
-
+        
         SECTION( "Multiplication-Assignment operator" ) {
             hA *= hB;
             REQUIRE( hA[0][0] == -43 );
@@ -2065,23 +2066,6 @@ TEST_CASE( "Hypercomplex: Polynomial lib test", "[unit]" ) {
             REQUIRE( hA[3][2] == -115 );
             REQUIRE( hA[3][3] == -134 );
             REQUIRE( hA[3][4] == -218 );
-            // test implicit type conversion
-            short int si = 3;
-            unsigned short int usi = 3;
-            int i = 3;
-            unsigned int ui = 3;
-            long int li = 3;
-            unsigned long int uli = 3;
-            long long int lli = 3;
-            unsigned long long int ulli = 3;
-            REQUIRE_NOTHROW(hA ^= si);
-            REQUIRE_NOTHROW(hA ^= usi);
-            REQUIRE_NOTHROW(hA ^= i);
-            REQUIRE_NOTHROW(hA ^= ui);
-            REQUIRE_NOTHROW(hA ^= li);
-            REQUIRE_NOTHROW(hA ^= uli);
-            REQUIRE_NOTHROW(hA ^= lli);
-            REQUIRE_NOTHROW(hA ^= ulli);
         }
 
         SECTION( "Output stream operator" ) {
