### 2023-04-15 AngryMaciek/hypercomplex#35 @AngryMaciek
Alright, I have added `#define CATCH_CONFIG_NO_POSIX_SIGNALS` as they suggest in their issue; on my machine and in the ci everything is OK; hopefully this fixes it for you too.
Warnings are OK; I am testing self-assignment operator in these exact lines.
