### 2023-04-15 AngryMaciek/hypercomplex#35 @AngryMaciek
Fixed on branch `revision` : could you please confirm?
