### 2023-04-15 AngryMaciek/hypercomplex#36 @AngryMaciek
I have added the code you specified above [as a separate test case](https://github.com/AngryMaciek/hypercomplex/blob/09cf984ac30473e499a35d823649c79dfb10970b/.test/unit/test.cpp#L1263) on a new branch `revision`; please see the [ci logs](https://github.com/AngryMaciek/hypercomplex/actions/runs/4707559851/jobs/8349485974?pr=37#step:8:104).
We get the same value as in the [docs](https://angrymaciek.github.io/hypercomplex/html/index.html).
Unfortunately we cannot reproduce the error; could this be specific to your machine?

I found the following information on the internet:

> More specifically, the assertion checks that the precision parameter p is greater than or equal to 1 and less than or equal to a maximum value determined by the size of the data type used to store the precision. This maximum value is calculated based on the platform's mpfr_uprec_t type, which specifies the largest unsigned integer type available on the system. The 256 offset is added to ensure some additional safety margin.
>
> Therefore, the error message indicates that the program attempted to use an invalid precision value that is either too small (less than 1) or too large (exceeding the maximum precision that can be stored on the system).
