### 2023-04-20 AngryMaciek/hypercomplex#36 @vissarion
This is the info getting from `uname -a`
```
Linux 5.19.0-40-generic #41~22.04.1-Ubuntu SMP PREEMPT_DYNAMIC Fri Mar 31 16:00:14 UTC 2 x86_64 x86_64 x86_64 GNU/Linux
```
