### 2023-04-20 AngryMaciek/hypercomplex#36 @vissarion
The same issue appear with clang++ 16. Note that the mpfr version in ci is 4.0.2 while in my local setting is 4.1.0, so this could be one explanation.
