+++ .gitpod.yml
@@ -1,0 +1,41 @@
+# This configuration file was automatically generated by Gitpod.
+# Please adjust to your needs (https://www.gitpod.io/docs/config-gitpod-file).
+
+image:
+  file: .gitpod.dockerfile
+
+tasks:
+
+  - name: main
+
+    # adjust PS1 + activate dev. env.
+    command: |
+      conda config --set changeps1 False
+      parse_git_branch() { git branch --show-current ; }
+      export PS1="\e[1;37m(\$(parse_git_branch)) \e[0;33m\w \e[1;37m$\e[0m "
+      echo "Ready to work!"
+
+ports:
+
+  - port: 8888
+    name: JupyterLab
+    description: Default port
+    visibility: private
+    onOpen: ignore
+
+github:
+  prebuilds:
+    # enable for the default branch
+    master: true
+    # enable for all branches in this repo
+    branches: false
+    # enable for pull requests coming from this repo
+    pullRequests: true
+    # enable for pull requests coming from forks
+    pullRequestsFromForks: false
+    # add a check to pull requests
+    addCheck: prevent-merge-on-error
+    # add a "Review in Gitpod" button as a comment to pull requests
+    addComment: false
+    # add a "Review in Gitpod" button to the pull request's description
+    addBadge: true
