### 2023-04-20 AngryMaciek/hypercomplex#36 @AngryMaciek
Are you sure you followed the exact installation instructions from the docs?

![Screenshot 2023-04-20 at 19 12 04](https://user-images.githubusercontent.com/19755308/233439673-235023ce-8c88-48d4-a9cb-11b6a32c318c.png)

On my personal mac machine I have a newer version and the tests pass too:
