### 2023-04-20 b7315ff
adjust test script
