### 2023-04-21 AngryMaciek/hypercomplex#36 @vissarion
OK, I found the issue, it seems that the above code is missing a `set_mpfr_precision()` statement. Then `get_mpfr_precision()` returns `0` and there is not enough precision to do the computation so mpfr assertion fails. In your test case most probably mpfr precision was set at some point before `get_mpfr_precision()` is executed and therefore the test was passing. I am closing this issue now.
