### 2023-04-21 AngryMaciek/hypercomplex#35 @vissarion
The compile time Catch2 error is now gone (in revision branch) but now I am getting a runtime `Segfault` in `TEST_CASE( "CD[256] | N = 257", "[local]" )`. The same if I comment out `TEST_CASE( "CD[256] | N = 257", "[local]" )` and uncomment `TEST_CASE( "CD[1024] | N = 1031", "[local]" )`. This is with both `g++-12` and `clang++-14` with `-O0` and without.
