### 2023-04-21 AngryMaciek/hypercomplex#35 @vissarion
For your reference this is the report from sanitizer
```
AddressSanitizer:DEADLYSIGNAL
=================================================================
==23354==ERROR: AddressSanitizer: stack-overflow on address 0x7ffd4db8d2f8 (pc 0x00000086a6f6 bp 0x7ffd4dd15c00 sp 0x7ffd4db8d2e0 T0)
    #0 0x86a6f6 in Hypercomplex<Polynomial<257u>, 256u> RingInverse<257u, 256u>(Hypercomplex<Polynomial<257u>, 256u> const&, long const&) hypercomplex/Hypercomplex.hpp:1532
    #1 0x7ee1bd in Hypercomplex<Polynomial<257u>, 256u> PUBLICKEY<257u, 256u>(Hypercomplex<Polynomial<257u>, 256u> const&, Hypercomplex<Polynomial<257u>, 256u> const&, long const&) hypercomplex/Hypercomplex.hpp:1574
    #2 0x618732 in ____C_A_T_C_H____T_E_S_T____152 /home/vissarion/workspace/hypercomplex/.test/unit/test.cpp:3176
    #3 0x4919a5 in Catch::TestInvokerAsFunction::invoke() const /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:14213
    #4 0x48c937 in Catch::TestCase::invoke() const /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:14056
    #5 0x46ceb7 in Catch::RunContext::invokeActiveTestCase() /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:12915
    #6 0x46b763 in Catch::RunContext::runCurrentTest(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> >&, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> >&) /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:12888
    #7 0x46106f in Catch::RunContext::runTest(Catch::TestCase const&) /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:12649
    #8 0x476ff4 in execute /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:13243
    #9 0x47ea91 in Catch::Session::runInternal() /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:13449
    #10 0x47d2e4 in Catch::Session::run() /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:13405
    #11 0x7ef9a5 in int Catch::Session::run<char>(int, char const* const*) /home/vissarion/workspace/hypercomplex/.test/unit/catch.hpp:13127
    #12 0x619b66 in main /home/vissarion/workspace/hypercomplex/.test/unit/test.cpp:3264
    #13 0x7f9d7f829d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
    #14 0x7f9d7f829e3f in __libc_start_main_impl ../csu/libc-start.c:392
    #15 0x40d4a4 in _start (/home/vissarion/workspace/hypercomplex/.test/unit/test+0x40d4a4)

SUMMARY: AddressSanitizer: stack-overflow hypercomplex/Hypercomplex.hpp:1532 in Hypercomplex<Polynomial<257u>, 256u> RingInverse<257u, 256u>(Hypercomplex<Polynomial<257u>, 256u> const&, long const&)
==23354==ABORTING
```
