### 2023-04-21 AngryMaciek/hypercomplex#35 @AngryMaciek
We are not getting any errors, neither in CI nor at the Gitpod test I yesterday added.
Could you please post the exact command you executed?
