### 2023-04-21 AngryMaciek/hypercomplex#35 @vissarion
I was running the commands as posted in the comment above https://github.com/AngryMaciek/hypercomplex/issues/35#issuecomment-1497334381
There is a `~[local]` argument missing, which when added results in all tests passing. So, this is fixed also on my side.
You can close the issue.
I think it would be better not to execute the "local" tests if the local argument is not given to somehow avoid segmentation faults if some user accidentally misses some argument. However, this is not a blocking issue.
