### 2023-04-21 AngryMaciek/hypercomplex#35 @vissarion
Also it would be helpful to add in the documentation detailed instructions on how to build and run the tests locally.
