### 2023-04-21 AngryMaciek/hypercomplex#35 @AngryMaciek
With the following information:

1. Test script passed without any issues while excluding just the last test (i.e. bigger input matrix)
2. The error msg: `SUMMARY: AddressSanitizer: stack-overflow...`

I am going to conclude that this is not a problem with the library itself, rather the stack size limit - exceeded by recursive calls of multiplications of big hypercomplex numbers. [Seems other ppl experienced this issue too](https://stackoverflow.com/questions/39691874/addresssanitizer-stack-overflow-at-address-error). Strangely enough _valgrind_ did not report any errors. Moreover, this `stack overflow` does not crash the program: encryption-decryption is still successful for all my test cases. I think this is enough information to consider this issue resolved.

Documentation updated: I added one more line how to run the unit test script to test the installation.

It is much easier to exclude test cases in _Catch2_ than to specify "to include all others by default". I have made sure that in all places I run unit test script I have `~[local]`, just to be sure no one gets confused how I call commands. The last test is commented out because it takes very long to run (~2h). However I tested it once on my machine, it passed.

Given both reviewers approved on the changes in the comments above I consider this thread finished. The issue will be closed on PR merge. For further revision please create new issues.
