### 2023-04-21 AngryMaciek/hypercomplex#38 @vissarion
[JOSS review] Some issues on references 
There are some DOIs missing

* Hoffstein, J., Pipher, J., & Silverman, J. H. (1998). NTRU: A ring-based public key cryptosys-100
tem. In J. P. Buhler (Ed.), Algorithmic number theory (pp. 267–288). Springer Berlin101
Heidelberg. ISBN: 978-3-540-69113-6

   https://doi.org/10.1007/BFb0054868

* Malekian, E., & Zakerolhosseini, A. (2010). OTRU: A non-associative and high speed public107
key cryptosystem. 2010 15th CSI International Symposium on Computer Architecture and108
Digital Systems, 83–90.

  https://doi.org/10.1109/CADS.2010.5623536

* Singh, S., Padhye, S., & Pal, A. (2021). Generalization of lattice-based cryptography on115
hypercomplex algebras. In P. Stănică, S. Gangopadhyay, & S. K. Debnath (Eds.), Security116
and privacy (pp. 67–79). Springer Singapore. ISBN: 978-981-33-6781-4

  https://doi.org/10.1007/978-981-33-6781-4_6

For Boost one can cite the current version (1.82 instead of 1.75). Also instead of citing Boost in general maybe cite specific libraries. E.g. qvm and math libraries (math implements https://www.boost.org/doc/libs/1_82_0/libs/math/doc/html/quaternions.html and https://www.boost.org/doc/libs/1_82_0/libs/math/doc/html/octonions.html )
