### 2023-04-21 #5272 @vissarion
@olexandr-konovalov I finished my review.

@AngryMaciek There is only one issue left to resolve, see https://github.com/AngryMaciek/hypercomplex/issues/38 regarding paper references. After resolving this I will check out the last point in my review checklist and recommend acceptance!
