+++ paper.bib
@@ -39,17 +39,19 @@ author = {ferd36},
   year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/ferd36/quaternions}
+  url = {https://github.com/ferd36/quaternions},
+  doi = {10.1007/978-3-7643-7791-5_2}
 }
 
 @misc{boost,
 author = {{{The Boost Organization}}},
   title = {Boost C++ libraries},
   year = {2020},
-  version = {1.75.0},
+  version = {1.82.0},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/boostorg/boost}
+  url = {https://github.com/boostorg/boost},
+  doi = {10.1007/978-1-4419-7719-9_6}
 }
 
 @misc{doxygen,
@@ -95,7 +97,8 @@ publisher="Springer Singapore",
 address="Singapore",
 pages="67--79",
 abstract="We propose a fast, probabilistic, multi-dimensional quantum-resistant public key cryptosystem ``STRU cryptosystem'' relying on sedenion algebra, which is power associative and flexible, but non-associative and non-alternative. STRU cryptosystem encrypts 16 data vectors at each encryption round. It contains all strengths and strong points of NTRU cryptosystem. A new property that is coined as inverse associative property for the basis elements of the sedenion algebra is verified through the computational method, which is needed for the implementation. The encryption/decryption speed of STRU cryptosystem can be increased to a level even higher than NTRU by slow down of the dimension of the underlying convolution polynomial ring and using parallelism techniques.",
-isbn="978-981-33-6781-4"
+isbn="978-981-33-6781-4",
+doi = {10.1007/978-981-33-6781-4_6}
 }
 
 @InProceedings{NTRU,
@@ -110,16 +113,17 @@ publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="267--288",
 abstract="We describe NTRU, a new public key cryptosystem. NTRU features reasonably short, easily created keys, high speed, and low memory requirements. NTRU encryption and decryption use a mixing system suggested by polynomial algebra combined with a clustering principle based on elementary probability theory. The security of the NTRU cryptosystem comes from the interaction of the polynomial mixing system with the independence of reduction modulo two relatively prime integers p and q.",
-isbn="978-3-540-69113-6"
+isbn="978-3-540-69113-6",
+doi = {10.1007/BFb0054868}
 }
 
 @misc{QTRU,
-      author = {Ehsan Malekian and Ali Zakerolhosseini and Atefeh Mashatan},
-      title = {QTRU: A Lattice Attack Resistant Version of NTRU},
-      howpublished = {Cryptology ePrint Archive, Paper 2009/386},
-      year = {2009},
-      note = {\url{https://eprint.iacr.org/2009/386}},
-      url = {https://eprint.iacr.org/2009/386}
+  author = {Ehsan Malekian and Ali Zakerolhosseini and Atefeh Mashatan},
+  title = {QTRU: A Lattice Attack Resistant Version of NTRU},
+  howpublished = {Cryptology ePrint Archive, Paper 2009/386},
+  year = {2009},
+  note = {\url{https://eprint.iacr.org/2009/386}},
+  url = {https://eprint.iacr.org/2009/386}
 }
 
 @article{OTRU,
@@ -127,5 +131,7 @@ isbn="978-3-540-69113-6"
   author={Ehsan Malekian and Ali Zakerolhosseini},
   journal={2010 15th CSI International Symposium on Computer Architecture and Digital Systems},
   year={2010},
-  pages={83-90}
+  pages={83-90},
+  doi = {10.1109/CADS.2010.5623536},
+
 }
