### 2023-04-21 AngryMaciek/hypercomplex#38 @AngryMaciek
Fixed on branch `revision`, I need to merge it now in order to the REVISION issue bot pick up the changes.
