### 2023-04-21 a8eee3f
refactor: JOSS revision, round 1 (#37)
