### 2023-04-30 #5272 @ludopulles
It's a cool repository and the code definitely adds useful tools to experiment with hypercomplex algebras.

However, I don't see this repository to have the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/review_criteria.html)  for the JOSS. Not to be disrespectful for the many hours put in, but there are no clear signals of effort in my opinion. Regarding some of the signals in the guidelines:
- The commit history length seems to be about 3 months with a large gap in between.
- The core 2 files in the hypercomplex directory contain <2000 loc and there is much boilerplate code. In fact, the specializations to mpfr_t and Polynomial can be avoided:
  - There are multiple wrappers for mpfr in C++, cf. https://www.mpfr.org/ § Interfaces for MPFR. Users can use their own interface and this makes this specification useless.
  - The specialization to Hypercomplex<Polynomial> seems to be exactly the same as the generic instance, except that the "norm" function is changed to the square (algebraic) norm, something which the generic case could also handle.
- Lastly, although the papers (QTRU, OTRU, STRU) point out there is exciting research done on hypercomplex algebras, these 3 papers were of theoretic kind and didn't have much of experiments. It seems the number of interested people in the crypto community is limited, and thus I would say with the current state it is unlikely someone will cite this repository in their work. For machine learning, I did find a [paper using hypercomplex algebras](https://arxiv.org/abs/2205.13273) (released after the repository's initial code of 2021) but this used an own implementation. As I am not an ML expert, I can't say how likely this repository is to be cited by ML researchers.

Still, the repository is a nice tool to play around with the hypercomplex algebras. It could still be useful to perhaps integrate this into the Boost library in a way that it is compatible with or supersedes the Quaternion and Octonion implementations.
