### 2023-05-01 #5272 @AngryMaciek
Dear @ludopulles,

To my understanding a "substantial scholarly effort" could be interpreted as:

1. a total number of work-hours the authors put into the project as a whole, as in _their effort_
2. a qualifiable (substantial) value the projects adds to a particular research field, by ex: introducing novel approaches or methods that could be of use to others.

The following repository excels at both aspects.

> The commit history length seems to be about 3 months with a large gap in between.

This is a statement of fact, not an argument. Besides: number & timing of commits over time is not a reliable estimate of effort put into a project. _JOSS_ itself states in the reviewer guidelines:
> While some authors contribute openly and accrue a long and rich commit history before submitting, others may upload their software to GitHub shortly before submitting their JOSS paper.

As I have mentioned in my [other post](https://github.com/openjournals/joss-reviews/issues/5159#issuecomment-1433252665) this is a resubmission of my previous work, which I have now elevated to a higher level. For the initial three months I have developed a general framework to operate on arbitrary large hypercomplex numbers from the Cayley-Dickson algebras. During the previous revision we agreed that the library is in a sufficient technical state for _JOSS_ but lacks a clear real-world application, which I agreed to implement. We ended the revision, I privated the repository and went back to the development and refactoring.

I have spent last three years (on-and-off, of course) studying cryptography and investigating post-quantum methods; specifically: lattice-based cryptography which NTRU and its variants are a part of. Implementing a generalised version of these systems in high dimensional algebras is not a trivial task. Operations related to inverting a polynomial in a modular quotient ring as well as inverting a whole hypercomplex number constructed from such polynomials required an implementation from scratch. Eventually, due to the complexity of the whole procedure I have first developed the whole cryptographic part  in _SageMath_, as it supports some algebraic operations out of the box. Only after verifying the cryptosystems, the code was reimplemented to _C++_ and included into the library (a necessary step: _Sage_ calculations are too slow for such applications). These parts are reflected in the `Polynomial.hpp` file as well as `RingInverse` methods in the main class. The most important functionality of the whole library is provided by `PUBLICKEY`, `ENCRYPT` and `DECRYPT` functions which are very short and simple exactly because they make use of properly overloaded operators and other pre-defined functions. Such a modular structure allowed for careful unit testing at each single step (see [this file](https://github.com/AngryMaciek/hypercomplex/blob/master/.test/unit/test.cpp)) and achieving 100% code coverage. All this was very tedious, but absolutely necessary to ensure mathematical correctness for calculations which could not possibly be traced any other way (ex: multiplication in a $1024$-dimensional algebra of polynomials modulo $x^{37}-1$).

I believe that the whole general framework (which I already had three years ago) combined with the design and manual implementation of everything related to quotient rings in _C++_, all tests as well as examples included in the manuscript (but also [see here](https://github.com/AngryMaciek/hypercomplex/blob/master/.test/docs/test.cpp)) which present novel cryptographic results (more on that below) are enough effort for a publication in the Journal of Open Source Software.

> The core 2 files in the hypercomplex directory contain <2000 loc and there is much boilerplate code

Please refer to the metrics as [reported by the whedon bot](https://github.com/openjournals/joss-reviews/issues/5272#issuecomment-1476172257), that is the official `cloc` configuration we should stick to. Granted: the repository contains _Catch2_ framework which should be excluded (that is: `C/C++ Header | 13617`, manually checked). Therefore I present a corrected table below:

```
Language                     files          blank        comment           code
C/C++ Header                                                           	   1159
C++                              4            134            246           3293
Markdown                         8             98              0            331
YAML                             7             74              0            273
TeX                              1             11              0            120
XML                              1              2              4            101
make                             1             18             32             40
HTML                             1              0              0              6
SUM:                            	                                   5323
```

That is ~5k  pure code lines in the repository. The core files are indeed <2k, however, it is not only the core that matters. This is a Journal of **Open Source Software** after all, and that is why other files with mechanisms ensuring proper software engineering standards (more on them later) should be considered just as well. I presume the _whedon_ bot is aware of that and that's why the reporting is configured the way it as (above). On top of that, if I recall correctly, previous _JOSS_ guidelines stated that submissions below 1k should be flagged for inspection (not rejected, flagged). With that information I believe there is no ground to argue the validity of this submission based on the line count.

> -   There are multiple wrappers for mpfr in C++, cf.  [https://www.mpfr.org/](https://www.mpfr.org/)  § Interfaces for MPFR. Users can use their own interface and this makes this specification useless.

_MPFR_ is a C library which operates on its own specific `struct`ures representing numbers of arbitrary precision. Almost all of operations specified in the `template Hypercomplex` class differ in their implementation one cannot manipulate _MPFR_ variables just as any others (use of built-in methods like: `mpfr_set_zero`, `mpfr_init2`, `mpfr_add` is required). In order to construct hypercomplex numbers with arbitrary-high precision this class specialisation is absolutely necessary.
Yes, there are existing _C++_ interfaces, but please notice what they say at the official website you have linked to:
>The following C++ interfaces for MPFR, _very different_ in their design (and in particular, in the strategies for intermediate precisions, so that they can yield different results), are available:

 - I do not know how these interfaces were tested (if tested, at all) and how they behave in all possible situations. You cannot possibly expect me to check multiple other available libraries and compare their results. It is much more clean that I provide a stable, reliable class specialisation myself.
 - Any bugs with precision handling on the interface side would lead to incorrect calculation with `Hypercomplex`. And most users would not question the interface, they would probably conclude that `Hypercomplex` is crap and move on. I'd prefer to avoid such situations.
 - I do not know if they (or: which of those) implemented all functionalities used by my overloaded operators for hypercomplex numbers. Any missing pieces would lead to errors. Again: it is better to provide a stable and reliable solution based on the _MPFR_ itself rather then rely on an interface.

> The specialization to Hypercomplex seems to be exactly the same as the generic instance, except that the "norm" function is changed to the square (algebraic) norm, something which the generic case could also handle.

I understand you mean `Polynomial` class here. Yes, here you are right: this specialisation is almost identical to the generic version with the exception that I need to explicitly forbid `e^H`, `norm` and `/` operators. But on the other hand I needed to add `CenterLift` and `RingInverse` and `norm2` functions which would operate specifically on `Hypercomplex<Polynomial>` instances and not on any others. Because of these minor adjustments I decided it would be cleaner to write out a full class specialisation in this case too. That is my personal choice.

> Lastly, although the papers (QTRU, OTRU, STRU) point out there is exciting research done on hypercomplex algebras, these 3 papers were of theoretic kind and didn't have much of experiments.

Exactly. This is probably _the strongest point in advantage_ of the following library. Publications which present novel cryptographic findings are purely theoretical. Meaning - there is no way for others to re-analyse, expand on or validate one's claims. There are just separate groups of people developing their own solutions and presenting just the final, usually unverifiable methods. This is the exact opposite of Open Science / Open Source. I believe that we ( as a community of researchers investigating cryptography) could greatly benefit from a high-quality reference point for lattice-based cryptosystems based on NTRU variants.

What is more, this is not just a hypothetical possibility. Errors really started to arise while I looked through the publications closely. In the current cryptographic literature there are two relatively recent papers which refer to STRU: _Thakur and Tripathi_ and _Singh et al._ The latter (more recent) states the following:
> There is an article, namely, “STRU: A Non-Alternative and Multi-dimensional Public Key Cryptosystem” given by Thakur and Tripathi in 2017 [15]. They proposed an STRU cryptosystem based on sedenions, but in the decryption process, they used associativity directly. Since we know that sedenions are non-associative, we cannot use associativity directly, as Thakur et al. did. So, their scheme does not follow the sedenionic requirements.

Therefore claims of the former duo are incorrect. What is worse: it is quite possible the paper by _Singh et al._ contains errors too. Section 4: "Extending the property to polynomials" does not look right. I tried to reproduce their results with `Hypercomplex` but with no success. Finally, I asked a [question](https://crypto.stackexchange.com/questions/103544/ntruencrypt-fails-on-sedonion-algebra) on Cryptography Stack Exchange and my doubts were confirmed. **It seems that both already published papers are flawed.** This could possibly have been avoided if only the authors or reviewers were aware of another piece of research that provides a stable implementation which would help them verify the original claims.
In my humble opinion: **the above alone is a significant scholarly effort** (as in: qualifiable novel value introduced to the research field).

> It seems the number of interested people in the crypto community is limited, and thus I would say with the current state it is unlikely someone will cite this repository in their work.

I disagree. As soon as we are done here I plan to contact authors of all above mentioned publications (and possibly other researchers working on X-TRU variants) and present them this library. Hopefully it will aid them in their further work and let them validate their claims *before* submitting a paper. There are little more embarrassing feelings then admitting tu publish a mistake or worse: retracting a whole publication.
I don't really see what are the basis for the comment above.

> For machine learning, I did find a  [paper using hypercomplex algebras](https://arxiv.org/abs/2205.13273)  (released after the repository's initial code of 2021) but this used an own implementation. As I am not an ML expert, I can't say how likely this repository is to be cited by ML researchers.

The fact that these ML researchers needed to develop their own custom code for hypercomplex operations signals a lack of other available solutions (ex: a well-established library). _Hypercomplex_ aims to fill this gap so that others (ex: ML specialists) would not need to develop algebraic operations on their own but rather focus on their domain. That's the core principle of open source: we are all building great things from smaller building blocks provided by others. No need to everyone to develop custom solutions to the same problem over and over.

> It's a cool repository and the code definitely adds useful tools to experiment with hypercomplex algebras.
>
> Still, the repository is a nice tool to play around with the hypercomplex algebras. It could still be useful to perhaps integrate this into the Boost library in a way that it is compatible with or supersedes the Quaternion and Octonion implementations.

The following submission is not a "tool to play around" but rather a state-of-the-art _C++_ implementation of an advanced post-quantum family of cryptosystems. First of its kind, as there in no other alternative which would allow to construct cryptosystems on such large algebras. That is most possibly because up to this date **no one have ever** proved that the X-TRU systems are able to function for sedonions (and higher) under the condition of specific choice of parameters and the initial value of $F$. That is no trivial statement as higher algebras loose crucial properties (associativity, commutativity, alternativity) and a generalisation into arbitrary large structures could just as well turn out to be impossible. I consider this a substantial result for the whole field of cryptography. Moreover, this is not just a "theoretical" publication which other researchers could trust on don't; the following repository is **the first one** to provide a working implementation and examples of encryption/decryption of full images (Fig.1) - hard to argue with a code anyone can run for themselves. The importance of presented work is quite substantial. Both in terms of scientific software for the community of computational mathematicians as well as conclusions for cryptographers.

Finally, why _JOSS_ is a suitable journal for this work:
As much as I present some novel results this is still a _C++_ library brought to a very high standard. Realising the lack of any other similar solution to analyse/verify proposed mechanisms for lattice-based systems I decided to fill this gap and provide a reference point for other researchers in the field of cryptography. In my opinion it is more valuable for the community as a whole to have one solid and stable solution on which we all can rely on then to have one more "theoretical" paper presenting fomulas.
That is also why I ensured the repository sticks to the highest standards of scientific software engineering:
* CI testing via GitHub Actions mechanism (running on both macOS and Ubuntu), these include:
  * Unit tests for all operations via _Catch2_ framework resulting in 100% code coverage
  * Proper docstrings for all code elements leading to a complete documentation via the `doxygen` engine.
  * static code analysis, i.e. _C++_ code linting to ensure a proper code style.
  * raw-include and system-wide installation test for the library (obviously)
* Developers _Makefile_ to carry out the above tests with `make`, that is to lower the energy barrier for contributions.
* Automatic setup and configuration of a cloud-based development environment at Gitpod; again to simplify contributing to the repository.
* Contributing instructions, Code of Conduct and a License (obviously).

Taking into account all of the points above I consider it *quite brave* to state that the following submission lacks "substantial scholarly effort".

Following the _JOSS_ guidelines:
> The decision on scholarly effort is ultimately one made by JOSS editors.

Thus I ask for the opinions of @olexandr-konovalov and @danielskatz

Kind Regards,
Maciek
