### 2023-05-11 #5272 @danielskatz
As the track editor here, I believe this meets JOSS's requirement for substantially scholarly effort.
