### 2023-05-11 #5272 @AngryMaciek
Dear @ludopulles,
It would still help me a lot if I could include the work in my publication list.
Could you please resume your revision as soon as possible?
Kind Regards,
Maciek
