+++ .github/workflows/unit.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [macos-10.15, ubuntu-20.04]
+        os: [macos-12, ubuntu-20.04]
       fail-fast: false
     env:
       working-directory: .test/unit
