### 2023-05-12 d41c91d
ci: upgrade macOS version (#42)
