+++ .github/workflows/install.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [macos-10.15, ubuntu-20.04]
+        os: [macos-12, ubuntu-20.04]
     env:
       test-directory: .test/install
       lib-directory: "hypercomplex"
