### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek sorry for being slow with replies. Many thanks @ludopulles and @vissarion for the thorough review, and to @AngryMaciek for the detailed answer. I also agree that the paper justifies the  substantially scholarly effort requirement. What next?

1. When the reviewers are satisfied with the improvements, we ask that they confirm their recommendation to accept the submission. @ludopulles did this explicitly, but I can't see @vissarion confirming that - could you do this please? I see you completed checklist, but e.g. are you happy with the response on the issues you have submitted to https://github.com/AngryMaciek/hypercomplex/issues?

2. @AngryMaciek When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed - that seems to be the case), and archive it (I'd recommend Zenodo as explained [here](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content), and alternatives mentioned at https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance). Please do this and post the version number and archive DOI in this review issue.
