### 2023-05-15 #5272 @vissarion
@olexandr-konovalov thanks for the ping, I though my comment https://github.com/openjournals/joss-reviews/issues/5272#issuecomment-1517745199 was enough, but provably it was not, sorry for that.

I finished with my review, my checklist and all the issues raised during this process are resolved. Therefore, I recommend acceptance of this paper!
