### 2023-05-15 #5272 @AngryMaciek
Info for @olexandr-konovalov:

Version `2.0.1` : https://github.com/AngryMaciek/hypercomplex/releases/tag/v2.0.1
_zenodo_: https://zenodo.org/record/7937280
DOI: `10.5281/zenodo.7937280`
