### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek thank you! Ideally Zenodo record should say

> Related identifiers: Supplement to https://github.com/AngryMaciek/hypercomplex/tree/v2.0.1

instead of

> Related identifiers: Derived from https://github.com/AngryMaciek/hypercomplex

to point to a specific release of the software (you can check recently published JOSS papers to see examples of this). Did you use GitHub-Zenodo integration, or uploaded the archive manually? Perhaps you can update the newly created Zenodo record to point to the precise tag? Thanks!
