### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek Thanks - looks good to me. If you want to have a preview (e.g. like https://doi.org/10.5281/zenodo.7676801) you can upload `.zip` instead of `.tar.gz` though (when you set up automated GitHub-Zenodo integration, it does it for you, and will seamlessly work for all future releases).
