### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek I've done proofreading - please see https://github.com/AngryMaciek/hypercomplex/pull/44 for my suggestions. In addition, there are some notes about the references:
- ferd36: points to Springer, not to GitHub
- Hoppe: Cayley dickson -> Cayley Dickson (capitalisation to match the source, and Dickson is a surname
- The Boost Organization. (2020) says "In GitHub" but the URL goes to Springer. If that resource is the recommended citation by Boost authors, use it but then drop "In GitHub". If you want to cite both sources, it would be fine to use two references.
