+++ paper.md
@@ -27,14 +27,14 @@ This is a highly specialised software aimed mostly for computational mathematici
 
 # Key features
 
-- As a header-only *C++* template code it's greatest advantage is the combination of speed, generic programming and convenience for the end user. Open Source license together with template specialisation mechanism allows contributors to add-in support for custom objects, define specific functions and extend the scope of the library.
+- As a header-only *C++* template code, it's greatest advantage is the combination of speed, generic programming, and convenience for the end user. Open Source license together with template specialisation mechanism allows contributors to add support for custom objects, define specific functions and extend the scope of the library.
 - The most important specialisation, already included in the library itself, is the introduction of operations in hypercomplex algebras over truncated polynomial rings. These allow for many cryptographic applications as described in a dedicated section below. 
 - Another template class specialisation introduces the support for arbitrary high precision of calculations via GNU MPFR library [@fousse:inria-00070266], for which the operators have been overloaded such that all the instructions are carried out on specific data structures.
 - State of the art technology for software engineering:
   - CI/CD mechanism set up with GitHub Actions: automatic tests for library installation, source code inclusion, compilation and execution,
   - extensive unit testing with Catch2 framework [@catch2] alongside code coverage measurement uploaded to Codecov; current coverage: 100%,
   - source code linting with cpplint [@cpplint] - Google code style enforced,
-  - automatic documentation generation and hosting on GitHub Pages: build via Doxygen [@doxygen], publishing via Actions.
+  - automatic documentation generation and hosting on GitHub Pages: build via Doxygen [@doxygen], publishing via GitHub Actions.
 
 # Cryptographic applications
 
@@ -47,16 +47,16 @@ Every element of $\mathcal{R}$, $\mathcal{R}_p$, $\mathcal{R}_q$ may be writted 
 f = \sum_{i=0}^{N-1} f_i x_i \equiv [f_0, \ldots ,f_{N-1}]
 \end{equation}
 
-Addition operation $+$ refers to a regular element-wise addition of coefficients (modular for $\mathcal{R}_p$ and $\mathcal{R}_q$).
+where the addition operation $+$ refers to a regular element-wise addition of coefficients (modular for $\mathcal{R}_p$ and $\mathcal{R}_q$).
 Multiplication $\star$ within this structure is defined as:
 
 \begin{equation}\label{eq:ringmul}
 f \star g = \sum_{i=0}^k f_i g_{k-i} + \sum_{i=k+1}^{N-1} f_i g_{N+k-i}
 \end{equation}
 
-With a final reduction modulo $p$ or $q$ in the modular quotient rings.
+with a final reduction modulo $p$ or $q$ in the modular quotient rings.
 
-Based on the above let us pick an integer $\lambda \geq 0$ and define three corresponding algebras, generated by the Cayley-Dickson process:
+Based on the above, let us pick an integer $\lambda \geq 0$ and define three corresponding algebras, generated by the Cayley-Dickson process:
 
 \begin{equation}\label{eq:algebras}
 \begin{aligned}
@@ -66,10 +66,11 @@ Based on the above let us pick an integer $\lambda \geq 0$ and define three corr
 \end{aligned}
 \end{equation}
 
-Note that $\forall x\in \mathcal{A^\lambda}: x = (a, b) | a, b \in \mathcal{A^{\lambda-1}}$.
+where the ddition operation $+$ refers to ring addition defined above.  
 
-Addition operation $+$ refers to ring addition defined above.  
-Multiplication $\times$ is defined recursively based on conjugation operation $^*$ as below:
+Note that $\forall x\in \mathcal{A^\lambda}: x = (a, b)$, where $a, b \in \mathcal{A^{\lambda-1}}$.
+
[…]
