+++ paper.bib
@@ -26,7 +26,7 @@
 
 @misc{cd,
 author = {T. Hoppe},
-  title = {Cayley Dickson},
+  title = {Cayley-Dickson},
   year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
