+++ paper.bib
@@ -26,7 +26,7 @@
 
 @misc{cd,
 author = {T. Hoppe},
-  title = {Cayley Dickson},
+  title = {CayleyDickson},
   year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -35,23 +35,28 @@ author = {T. Hoppe},
 
 @misc{quaternions,
 author = {ferd36},
-  title = {Quaternions},
-  year = {last accessed: 16 Feb 2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/ferd36/quaternions},
-  doi = {10.1007/978-3-7643-7791-5_2}
+  title="Quaternions",
+  bookTitle="Quaternions, Clifford Algebras and Relativistic Physics",
+  year="2007",
+  publisher="Birkh{\"a}user Basel",
+  address="Basel",
+  pages="3--17",
+  isbn="978-3-7643-7791-5",
+  doi="10.1007/978-3-7643-7791-5_2",
+  url="https://doi.org/10.1007/978-3-7643-7791-5_2"
 }
 
 @misc{boost,
-author = {{{The Boost Organization}}},
-  title = {Boost C++ libraries},
-  year = {2020},
-  version = {1.82.0},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/boostorg/boost},
-  doi = {10.1007/978-1-4419-7719-9_6}
+author="Koranne, Sandeep",
+  title="Boost C++ Libraries",
+  bookTitle="Handbook of Open Source Tools",
+  year="2011",
+  publisher="Springer US",
+  address="Boston, MA",
+  pages="127--143",
+  isbn="978-1-4419-7719-9",
+  doi="10.1007/978-1-4419-7719-9_6",
+  url="https://doi.org/10.1007/978-1-4419-7719-9_6"
 }
 
 @misc{doxygen,
