### 2023-05-15 #5272 @AngryMaciek
@olexandr-konovalov : thank you for reading over the text in detail;
I have merged your branch "as was" and corrected the references according to the above; I could not get a proper capitalization for `dickson`, regardless if I used a space of a dash in the `bib` file it always came out parsed as lowercase (parsing engine bug?) In the end I wrote `CayleyDickson` as this is closest to the truth.
I have generated a next release `2.0.2` and a new version on _zenodo_:
https://doi.org/10.5281/zenodo.7938707
DOI: 10.5281/zenodo.7938707
