### 2023-05-15 5419913
docs: fix ref. capitalisation syntax
