### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek you need to use extra `{..}` to enforce capitalisation, e.g. `{Cayley {D}ickson}` or even `{{Cayley Dickson}}`- and some other entries in the bibliography would benefit from this too, otherwise it says `c++` instead of `C++`.
