+++ paper.bib
@@ -48,7 +48,7 @@ author = {ferd36},
 
 @misc{boost,
 author="Koranne, Sandeep",
-  title="Boost C++ Libraries",
+  title = {{Boost C++ Libraries}},  
   bookTitle="Handbook of Open Source Tools",
   year="2011",
   publisher="Springer US",
