### 2023-05-15 #5272 @olexandr-konovalov
@AngryMaciek excellent, thanks! Since the paper is part of the master branch of the repository, of course the cleanest way would be to make a new release with THIS version of the paper - would you be able to do this, and then I will proceed with the rest of the TODO list at https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance ?
