### 2023-05-15 #5272 @danielskatz
Technically, we don't need the paper to be in the release at all...  Many authors put it in a branch that isn't part of the release. We archive the paper separately once it is published.

but having it in the release is fine as well.
