### 2023-05-15 #5272 @olexandr-konovalov
True - I assumed it's author's choice, but the [instructions](https://joss.readthedocs.io/en/latest/editing.html#overview-of-editorial-process) allow both options. I'd put it in a separate branch if I would be writing a paper.

Anyhow, perhaps for now this is perfect, we don't need yet another release! In the future, you would be able to move it away from your master branch. Thanks @AngryMaciek for responding to all suggestions quickly!
