### 2023-05-15 #5272 @danielskatz
As track editor now taking over for the final processing, I'm suggesting a bunch of changes in https://github.com/AngryMaciek/hypercomplex/pull/46

Also, I don't understand "ferd36" as a an author - should this be Patrick R. Girard, the author of the overall book?
