+++ paper.bib
@@ -34,7 +34,7 @@ author = {T. Hoppe},
 }
 
 @misc{quaternions,
-author = {ferd36},
+author = {Patrick R. Girard},
   title="Quaternions",
   bookTitle="Quaternions, Clifford Algebras and Relativistic Physics",
   year="2007",
