### 2023-05-15 b25ee5a
docs: editorial changes in paper and bib (#46)
