+++ paper.bib
@@ -10,7 +10,7 @@
 }
 
 @techreport{fousse:inria-00070266,
-  TITLE = {{MPFR: A Multiple-Precision Binary Floating-Point Library With Correct Rounding}},
+  TITLE = {{MPFR}: A Multiple-Precision Binary Floating-Point Library With Correct Rounding},
   AUTHOR = {Fousse, Laurent and Hanrot, Guillaume and Lef{\`e}vre, Vincent and P{\'e}lissier, Patrick and Zimmermann, Paul},
   URL = {https://hal.inria.fr/inria-00070266},
   TYPE = {Research Report},
@@ -34,7 +34,7 @@ author = {T. Hoppe},
 }
 
 @misc{quaternions,
-author = {ferd36},
+author = {Patrick R. Girard},
   title="Quaternions",
   bookTitle="Quaternions, Clifford Algebras and Relativistic Physics",
   year="2007",
@@ -48,7 +48,7 @@ author = {ferd36},
 
 @misc{boost,
 author="Koranne, Sandeep",
-  title = {{Boost C++ Libraries}},  
+  title = {Boost {C}++ Libraries},  
   bookTitle="Handbook of Open Source Tools",
   year="2011",
   publisher="Springer US",
@@ -60,7 +60,7 @@ author="Koranne, Sandeep",
 }
 
 @misc{doxygen,
-author = {{{van Heesch, D.}}},
+author = {van Heesch, D.},
   title = {Doxygen},
   year = {2021},
   version = {1.9.1},
@@ -70,8 +70,8 @@ author = {{{van Heesch, D.}}},
 }
 
 @misc{cpplint,
-author = {{{Google Inc.}}},
-  title = {cpplint},
+author = {{Google Inc.}},
+  title = {{cpplint}},
   year = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
