### 2023-05-15 #5272 @AngryMaciek
Done, but give me a second please, page formatting is off, I will adjust the document once again.
