+++ paper.bib
@@ -60,7 +60,7 @@ author="Koranne, Sandeep",
 }
 
 @misc{doxygen,
-author = {van Heesch, D.},
+author = {{van Heesch}, D.},
   title = {Doxygen},
   year = {2021},
   version = {1.9.1},
