+++ paper.bib
@@ -27,7 +27,7 @@
 @misc{cd,
 author = {T. Hoppe},
   title = {{Cayley Dickson}},
-  year = {last accessed: 16 Feb 2021},
+  note = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/thoppe/Cayley-Dickson}
@@ -72,7 +72,7 @@ author = {van Heesch, D.},
 @misc{cpplint,
 author = {{Google Inc.}},
   title = {{cpplint}},
-  year = {last accessed: 16 Feb 2021},
+  note = {last accessed: 16 Feb 2021},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/google/styleguide}
