### 2023-05-15 #5272 @danielskatz
@AngryMaciek - sorry, one more fix that shouldn't be needed, but appears to be, plus a couple of things I missed: https://github.com/AngryMaciek/hypercomplex/pull/48
