### 2023-05-15 #5272 @AngryMaciek
Thank you @danielskatz @olexandr-konovalov @vissarion @ludopulles  for your time and effort to improve this work!

I am unimaginably happy to finalise this. What started as late-evening pet project years ago now evolved into original research in cryptography with state-of-the-art software engineering mechanisms. The scope grew waaaay beyond what I initially planed but now, as I pushed through all obstacles, I am really proud of it; I think it's the most challenging piece of math I've ever done on my own :)

All the best to everyone!
