### 2023-05-22 #5272 @AngryMaciek
Dear @olexandr-konovalov and @danielskatz,

I have presented this work to my colleagues and they have unfortunately pointed out numerous embarrassing typos in mathematical notations which we have all missed here:

1. Multiplicative group of integers modulo $p$ and $q$ should be: $Z/pZ$ and $Z/qZ$
2. Lower suffix in polynomial of $x$ should be an upper suffix as a power operation: $x^i$
3. Multiplication definition in its current form denotes only a coefficient for $x^k$, that is $[f \star g]_k$; full multiplication formula requires a summation over all $k$ values and a $x^k$ term

I have denoted the corrections below:

---

I understand this is a problem and I apologise for the situation but could we please adjust the publication as mentioned in the previous post (pointing to the JOSS guidelines)?

> @AngryMaciek [here](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance) it is - please check the final proof. As said here, changing papers after the publication is very tedioius.

Kind Regards,
