### 2023-05-22 #5272 @AngryMaciek
@danielskatz - I have corrected the typos and clarified formula (2): current indices reflect formula (1) better so that they are visually closer; current `master` branch contains a fixed version of `paper.md` and that is the only file that was changed in the PR (https://github.com/AngryMaciek/hypercomplex/pull/54)
